+++ bot.py
@@ -6,217 +6,438 @@ from telegram.ext import (
 from telegram import Update, InlineKeyboardButton, InlineKeyboardMarkup
 from telegram.constants import ParseMode
 import traceback, logging
+from datetime import datetime as dt, timedelta
 import json as js
 from usage import Usage
 from messages import Messages
 from pcs import PCS
 from chain import Chain
+#from music import Music
+from users import Users
+from util import Util
+from warnings import filterwarnings
+from telegram.warnings import PTBUserWarning
+
+#Filterint warnings about conversation handlers configuration...
+filterwarnings(action="ignore", message=r".*CallbackQueryHandler", category=PTBUserWarning)
 
 print("Starting musiCal Bot...", end="\n")
 config = js.load(open("config.json")) #The configuration file (token included)
 en_users = set() #In this set the bot store ids from users who prefer to speak in English
-us = Usage("usage.csv", "errors.csv") #The class to work with usage data...
+us = Usage("data/usage.csv", "data/errors.csv") #The class to work with usage data...
 msg = Messages() #The class to build content of text messages...
 pcs = PCS() #A class to analyze pitch class sets...
-PCS_S, CHAIN_S, MATRIX_1, MATRIX_2, ERROR_1, ERROR_2 = range(6) #The general conversation states...
+users = Users("data/users/") #A class to save user's configuration...
+ut = Util() #Some useful functions...
+#mus = Music() #A class to create experimental music...
+START_AN, START_TS, PCS_S, CHAIN_S, RANDOM_S, RANDOM_D, RANDOM_B, RANDOM_C, ERROR_1, ERROR_2, ADMIN = range(11) #The general conversation states...
+#CP_T, CP_M, CP_D = range(3) #The music conversation states...
 
 #Welcome message for people who start the bot...
 async def start(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
-	id = update.effective_chat.id
-	logging.info(str(hide_id(id)) + " started the bot...")
+	chat_id = update.effective_chat.id
+	logging.info(str(hide_id(chat_id)) + " started the bot...")
 	us.add_start()
-	await context.bot.send_message(chat_id=id, text=msg.get_message("hello", get_language(id)), parse_mode=ParseMode.HTML)
+	await context.bot.send_message(chat_id=chat_id, text=msg.get_message("start_1", get_language(context)), parse_mode=ParseMode.HTML)
+	await context.bot.send_message(chat_id=chat_id, text=msg.get_message("start_2", get_language(context)), parse_mode=ParseMode.HTML)
+	return START_AN
+
+#Saving the user's artistic name...
+async def save_artistic_name(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
+	chat_id = update.effective_chat.id
+	m = update.message.text
+	context.chat_data["name"] = m
+	await context.bot.send_message(chat_id=chat_id, text=msg.get_message("start_3", get_language(context)), parse_mode=ParseMode.HTML)
+	return START_TS
+
+#Saving the user's prefered time signature...
+async def save_prefered_time_signature(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
+	chat_id = update.effective_chat.id
+	m = update.message.text
+	if msg.is_time_signature(m):
+		context.chat_data["t_signature"] = m
+		users.save_user_data(chat_id, context.chat_data)
+		await context.bot.send_message(chat_id=chat_id, text=msg.get_message("start_5", get_language(context)), parse_mode=ParseMode.HTML)
+		return ConversationHandler.END
+	else:
+		await context.bot.send_message(chat_id=chat_id, text=msg.get_message("start_4", get_language(context)), parse_mode=ParseMode.HTML)
+		return START_TS
 
 #Starting a pitch class set analysis session...
 async def trigger_pcs(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-	id = update.effective_chat.id
-	logging.info(str(hide_id(id)) + " starts pcs conversation...")
-	await context.bot.send_message(chat_id=id, text=msg.get_conversation_start(get_language(id)), parse_mode=ParseMode.HTML)
-	await context.bot.send_message(chat_id=id, text=msg.get_message("pcs_1", get_language(id)), parse_mode=ParseMode.HTML)
+	chat_id = update.effective_chat.id
+	logging.info(str(hide_id(chat_id)) + " starts pcs conversation...")
+	await context.bot.send_message(chat_id=chat_id, text=msg.get_conversation_start(get_language(context)), parse_mode=ParseMode.HTML)
+	await context.bot.send_message(chat_id=chat_id, text=msg.get_message("pcs_1", get_language(context)), parse_mode=ParseMode.HTML)
 	return PCS_S
 
 #Analyzing pitch class sets sent by the user...
 async def get_pcs_info(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-	id = update.effective_chat.id
+	chat_id = update.effective_chat.id
 	text = update.message.text
 	try:
 		cardinality, ordinal, interval, is_inverted, z_pair, states, ordered_form, prime_form = pcs.get_set_info(text)
 		vector = pcs.interval_vector(ordered_form)
 		if not ordinal == None:
-			m = msg.build_pcs_message(cardinality, ordinal, interval, is_inverted, z_pair, states, ordered_form, prime_form, vector, get_language(id))
+			m = msg.build_pcs_message(cardinality, ordinal, interval, is_inverted, z_pair, states, ordered_form, prime_form, vector, get_language(context))
 			us.add_pcs(0)
-			await context.bot.send_message(chat_id=id, text=m, parse_mode=ParseMode.HTML)
+			await context.bot.send_message(chat_id=chat_id, text=m, parse_mode=ParseMode.HTML)
 		else:
 			us.add_pcs(1)
-			await context.bot.send_message(chat_id=id, text=msg.get_message("pcs_2", get_language(id)), parse_mode=ParseMode.HTML)
+			await context.bot.send_message(chat_id=chat_id, text=msg.get_message("pcs_2", get_language(context)), parse_mode=ParseMode.HTML)
 	except:
 		us.add_pcs(2)
-		await context.bot.send_message(chat_id=id, text=msg.get_message("pcs_3", get_language(id)), parse_mode=ParseMode.HTML)
+		await context.bot.send_message(chat_id=chat_id, text=msg.get_message("pcs_3", get_language(context)), parse_mode=ParseMode.HTML)
 	return PCS_S
 
 #Starting a constant pitch class set sequence creation session...
 async def trigger_chain(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-	id = update.effective_chat.id
-	logging.info(str(hide_id(id)) + " starts chain conversation...")
-	await context.bot.send_message(chat_id=id, text=msg.get_conversation_start(get_language(id)), parse_mode=ParseMode.HTML)
-	await context.bot.send_message(chat_id=id, text=msg.get_message("chain_1", get_language(id)), parse_mode=ParseMode.HTML)
+	chat_id = update.effective_chat.id
+	logging.info(str(hide_id(chat_id)) + " starts chain conversation...")
+	await context.bot.send_message(chat_id=chat_id, text=msg.get_conversation_start(get_language(context)), parse_mode=ParseMode.HTML)
+	await context.bot.send_message(chat_id=chat_id, text=msg.get_message("chain_1", get_language(context)), parse_mode=ParseMode.HTML)
 	return CHAIN_S
 
 #Creating a new constant pitch class set notes sequence...
 async def get_new_chain(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-	id = update.effective_chat.id
+	chat_id = update.effective_chat.id
 	text = update.message.text
 	try:
 		the_chain = Chain(pcs, text, 7, 14, 3)
-		m = msg.build_chain_message(the_chain, get_language(id))
+		m = msg.build_chain_message(the_chain, get_language(context))
 		us.add_chain(0)
-		await context.bot.send_message(chat_id=id, text=m, parse_mode=ParseMode.HTML)
+		await context.bot.send_message(chat_id=chat_id, text=m, parse_mode=ParseMode.HTML)
 	except:
 		us.add_chain(1)
-		await context.bot.send_message(chat_id=id, text=msg.get_message("chain_2", get_language(id)), parse_mode=ParseMode.HTML)
+		await context.bot.send_message(chat_id=chat_id, text=msg.get_message("chain_2", get_language(context)), parse_mode=ParseMode.HTML)
 	return CHAIN_S
+
+#Starting a random functions session...
+async def trigger_random(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
+	chat_id = update.effective_chat.id
+	logging.info(str(hide_id(chat_id)) + " starts random conversation...")
+	await context.bot.send_message(chat_id=chat_id, text=msg.get_conversation_start(get_language(context)), parse_mode=ParseMode.HTML)
+	await context.bot.send_message(chat_id=chat_id, text=msg.get_message("random_1", get_language(context)), parse_mode=ParseMode.HTML)
+	return RANDOM_S
+
+#Saving the sequence size preference...
+async def set_random_size(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
+	chat_id = update.effective_chat.id
+	text = update.message.text
+	try:
+		size = int(text)
+		context.chat_data["random_size"] = size
+		await context.bot.send_message(chat_id=chat_id, text=msg.get_message("random_2", get_language(context)), parse_mode=ParseMode.HTML)
+		return RANDOM_D
+	except:
+		us.add_random(1)
+		await context.bot.send_message(chat_id=chat_id, text=msg.get_message("random_5", get_language(context)), parse_mode=ParseMode.HTML)
+		return RANDOM_S
+
+#Saving the dice size preference...
+async def set_dice_size(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
+	chat_id = update.effective_chat.id
+	text = update.message.text
+	try:
+		size = int(text)
+		context.chat_data["random_dice"] = size
+		keyboard = random_keyboard(get_language(context))
+		reply = InlineKeyboardMarkup(keyboard)
+		await context.bot.send_message(chat_id=chat_id, text=msg.get_message("random_3", get_language(context)), reply_markup=reply, parse_mode=ParseMode.HTML)
+		return RANDOM_B
+	except:
+		us.add_random(1)
+		await context.bot.send_message(chat_id=chat_id, text=msg.get_message("random_5", get_language(context)), parse_mode=ParseMode.HTML)
+		return RANDOM_D
+
+#Building the random keyboard...
+def random_keyboard(language: int):
+	tags = [] 
+	if language == 0:
+		tags = ["Un dado", "Menor de dos", "Suma de dos", "Suma de tres", "Con lista"]
+	elif language == 1:
+		tags = ["A dice", "Minor of two", "Sum of two", "Sum of three", "With list"]
+	keyboard = [[InlineKeyboardButton(text=tags[0], callback_data="r_0"), InlineKeyboardButton(text=tags[1], callback_data="r_1")],
+							[InlineKeyboardButton(text=tags[2], callback_data="r_2"), InlineKeyboardButton(text=tags[3], callback_data="r_3")],
+							[InlineKeyboardButton(text=tags[4], callback_data="r_4")]]
+	return keyboard
+
+#Shuffling a user's message...
+async def shuffle_message(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
+	chat_id = update.effective_chat.id
+	text = update.message.text
+	m = ut.shuffle_string(text)
+	await context.bot.send_message(chat_id=chat_id, text=m, parse_mode=ParseMode.HTML)
+	return RANDOM_B
+
+#Random message with user's list...
+async def get_controlled_random(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
+	chat_id = update.effective_chat.id
+	text = update.message.text
+	m = ut.controlled_random(context.chat_data["random_size"], text.split(" "))
+	await context.bot.send_message(chat_id=chat_id, text=m, parse_mode=ParseMode.HTML)
+	return RANDOM_B
+
+#Handling random conversation clicks...
+async def random_conversation_button_click(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
+	chat_id = update.effective_chat.id
+	query = update.callback_query
+	await query.answer()
+	us.add_random(0)
+	selection = int(query.data.split("_")[1])
+	if selection == 0:
+		m = msg.sequence_to_string(ut.dice(context.chat_data["random_size"], context.chat_data["random_dice"]))
+		await context.bot.send_message(chat_id=chat_id, text=m, parse_mode=ParseMode.HTML)
+		return RANDOM_B
+	elif selection == 1:
+		m = msg.sequence_to_string(ut.minor_of_a_pair(context.chat_data["random_size"], context.chat_data["random_dice"]))
+		await context.bot.send_message(chat_id=chat_id, text=m, parse_mode=ParseMode.HTML)
+		return RANDOM_B
+	elif selection == 2:
+		m = msg.sequence_to_string(ut.sum_of_dice(context.chat_data["random_size"], context.chat_data["random_dice"], 2))
+		await context.bot.send_message(chat_id=chat_id, text=m, parse_mode=ParseMode.HTML)
+		return RANDOM_B
+	elif selection == 3:
+		m = msg.sequence_to_string(ut.sum_of_dice(context.chat_data["random_size"], context.chat_data["random_dice"], 3))
+		await context.bot.send_message(chat_id=chat_id, text=m, parse_mode=ParseMode.HTML)
+		return RANDOM_B
+	elif selection == 4:
+		await context.bot.send_message(chat_id=chat_id, text=msg.get_message("random_4", get_language(context)), parse_mode=ParseMode.HTML)
+		return RANDOM_C
+
+#Starting a music creation session...
+#async def trigger_mus(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
+#	chat_id = update.effective_chat.id
+#	logging.info(str(hide_id(chat_id)) + " starts a music creation session...")
+#	await context.bot.send_message(chat_id=chat_id, text=msg.get_message("music_t", get_language(context)), parse_mode=ParseMode.HTML)
+#	return CP_T
+
+#Deciding a title for the composition...
+#async def get_mus_title(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
+#	chat_id = update.effective_chat.id
+#	m = update.message.text
+#	context.chat_data["music_t"] = m
+#	await context.bot.send_message(chat_id=chat_id, text=msg.get_message("music_c_m", get_language(context)), parse_mode=ParseMode.HTML)
+#	return CP_M
+
+#Deciding a title for the composition...
+#async def get_mus_mode(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
+#	chat_id = update.effective_chat.id
+#	m = update.message.text
+#	if m == "1":
+#		mode = "random"
+#	elif m == "2":
+#		mode = "control"
+#	elif m == "3":
+#		mode = "markov"
+#	context.chat_data["music_m"] = mode
+#	await context.bot.send_message(chat_id=chat_id, text=msg.get_message("music_c_d", get_language(context)), parse_mode=ParseMode.HTML)
+#	return CP_D
+
+#Creating and sending the counterpoint...
+#async def build_counterpoint(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
+#	chat_id = update.effective_chat.id
+#	data = update.message.text
+#	title = context.chat_data["music_t"]
+#	mode = context.chat_data["music_m"]
+#	if mode == "control":
+#		c = mus.control_counterpoint(title, "musiCalBot", data)
+#	elif mode = "markov":
+#		c = mus.markov_counterpoint(title, "musiCalBot")
+#	elif mode = "random":
+#		c = mus.random_counterpoint(title, "musiCalBot")
+#	filename = "temp/" + title + "_" + composer + ".mid"
+#	await context.bot.send_message(chat_id=chat_id, text=msg.get_message("music_c", get_language(context)), parse_mode=ParseMode.HTML)
+#	await context.bot.send
+#	return ConversationHandler.END
 
 #Starting an error report session...
 async def trigger_error_submit(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-	id = update.effective_chat.id
-	logging.info(str(hide_id(id)) + " wants to report an error...")
-	await context.bot.send_message(chat_id=id, text=msg.get_apology(get_language(id)), parse_mode=ParseMode.HTML)
-	await context.bot.send_message(chat_id=id, text=msg.get_message("submit_error_1", get_language(id)), parse_mode=ParseMode.HTML)
+	chat_id = update.effective_chat.id
+	logging.info(str(hide_id(chat_id)) + " wants to report an error...")
+	await context.bot.send_message(chat_id=chat_id, text=msg.get_apology(get_language(context)), parse_mode=ParseMode.HTML)
+	await context.bot.send_message(chat_id=chat_id, text=msg.get_message("submit_error_1", get_language(context)), parse_mode=ParseMode.HTML)
 	return ERROR_1
 
 #Saving error related command...
 async def report_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-	id = update.effective_chat.id
+	chat_id = update.effective_chat.id
 	m = update.message.text
 	context.chat_data["error_command"] = m
-	await context.bot.send_message(chat_id=id, text=msg.get_message("submit_error_2", get_language(id)), parse_mode=ParseMode.HTML)
+	await context.bot.send_message(chat_id=chat_id, text=msg.get_message("submit_error_2", get_language(context)), parse_mode=ParseMode.HTML)
 	return ERROR_2
 
 #Saving error description...
 async def report_error(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-	id = update.effective_chat.id
+	chat_id = update.effective_chat.id
 	m = context.chat_data["error_command"]
 	m2 = update.message.text
 	context.chat_data["error_description"] = m2
 	us.add_error_report()
-	us.save_error_report(m, m2, str(hide_id(id)))
+	us.save_error_report(m, m2, str(hide_id(chat_id)))
 	admin_msg = "Error reported:\n-command: " + m + "\n-description: " + m2
 	await context.bot.send_message(chat_id=config["admin_id"], text=admin_msg, parse_mode=ParseMode.HTML)
-	await context.bot.send_message(chat_id=id, text=msg.get_message("submit_error_3", get_language(id)), parse_mode=ParseMode.HTML)
+	await context.bot.send_message(chat_id=chat_id, text=msg.get_message("submit_error_3", get_language(context)), parse_mode=ParseMode.HTML)
 	return ConversationHandler.END
+
+#Starting admin session...
+async def trigger_admin(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
+	chat_id = update.effective_chat.id
+	password = update.message.text.split(" ")
+	if len(password) > 1 and password[1] == config["password"]:
+		us.add_admin()
+		context.chat_data["admin_t"] = dt.now()
+		if get_language(context) == 0:
+			keyboard = [[InlineKeyboardButton(text="Últimos datos", callback_data="a_0")]]
+		else:
+			keyboard = [[InlineKeyboardButton(text="Last data", callback_data="a_0")]]
+		reply = InlineKeyboardMarkup(keyboard)
+		await context.bot.send_message(chat_id=chat_id, text=msg.get_message("admin", get_language(context)), reply_markup=reply, parse_mode=ParseMode.HTML)
+		return ADMIN
+	else:
+		logging.info(hide_id(chat_id) + " wanted to start an admin session...")
+		await context.bot.send_message(chat_id=chat_id, text=msg.get_message("intruder", get_language(context)), parse_mode=ParseMode.HTML)
+		return ConversationHandler.END
 
 #Ending any convertation...
 async def end_conversation(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-	id = update.effective_chat.id
-	logging.info(str(hide_id(id)) + " endss a conversation...")
-	await context.bot.send_message(chat_id=id, text=msg.get_conversation_end(get_language(id)), parse_mode=ParseMode.HTML)
-	await context.bot.send_message(chat_id=id, text=msg.get_message("end_conversation", get_language(id)), parse_mode=ParseMode.HTML)
+	chat_id = update.effective_chat.id
+	logging.info(str(hide_id(chat_id)) + " endss a conversation...")
+	await context.bot.send_message(chat_id=chat_id, text=msg.get_conversation_end(get_language(context)), parse_mode=ParseMode.HTML)
+	await context.bot.send_message(chat_id=chat_id, text=msg.get_message("end_conversation", get_language(context)), parse_mode=ParseMode.HTML)
 	return ConversationHandler.END
+
+#Printing privacy command...
+async def print_privacy(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
+	chat_id = update.effective_chat.id
+	logging.info(str(hide_id(chat_id)) + " checked privacy policy...")
+	await context.bot.send_message(chat_id=chat_id, text=msg.get_message("privacy", get_language(context)), parse_mode=ParseMode.HTML)
 
 #Sending a help message...
 async def print_help(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
-	id = update.effective_chat.id
-	logging.info(str(hide_id(id)) + " asked for help...")
+	chat_id = update.effective_chat.id
+	logging.info(str(hide_id(chat_id)) + " asked for help...")
 	us.add_help()
-	m, m2 = msg.build_help_message(get_language(id))
-	await context.bot.send_message(chat_id=id, text=m, parse_mode=ParseMode.HTML)
-	await context.bot.send_message(chat_id=id, text=m2, parse_mode=ParseMode.HTML)
+	m, m2 = msg.build_help_message(get_language(context))
+	await context.bot.send_message(chat_id=chat_id, text=m, parse_mode=ParseMode.HTML)
+	await context.bot.send_message(chat_id=chat_id, text=m2, parse_mode=ParseMode.HTML)
 
 #Checking which language to use with the actual user...
-def get_language(id):
-	if id in en_users:
-		return 1
+def get_language(context: ContextTypes.DEFAULT_TYPE) -> None:
+	if "language" in context.chat_data:
+		return context.chat_data["language"]
 	else:
 		return 0
 
-#A commando to allow a user decide which language to use...
+#A command to allow a user decide which language to use...
 async def select_language(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
-	id = update.effective_chat.id
-	logging.info(str(hide_id(id)) + " will set language...")
+	chat_id = update.effective_chat.id
+	logging.info(str(hide_id(chat_id)) + " will set language...")
 	keyboard = [[InlineKeyboardButton(text="Español", callback_data="l_0"),
 				InlineKeyboardButton(text="English", callback_data="l_1")]]
 	reply = InlineKeyboardMarkup(keyboard)
-	await context.bot.send_message(chat_id=id, text=msg.get_message("language", get_language(id)), reply_markup=reply, parse_mode=ParseMode.HTML)
+	await context.bot.send_message(chat_id=chat_id, text=msg.get_message("language", get_language(context)), reply_markup=reply, parse_mode=ParseMode.HTML)
 
 #Setting language configuration for actual user...
 async def set_language(update: Update, context: ContextTypes.DEFAULT_TYPE, query) -> None:
-	id = update.effective_chat.id
+	chat_id = update.effective_chat.id
 	if query == "l_1":
-		logging.info("English is the language selected by " + str(hide_id(id)))
-		en_users.add(id)
+		logging.info("English is the language selected by " + str(hide_id(chat_id)))
+		context.chat_data["language"] = 1
 		us.add_language(1)
-		await context.bot.send_message(chat_id=id, text=msg.get_message("language2", get_language(id)), parse_mode=ParseMode.HTML)
+		await context.bot.send_message(chat_id=chat_id, text=msg.get_message("language2", get_language(context)), parse_mode=ParseMode.HTML)
 	else:
-		logging.info("Spanish is the language selected by " + str(hide_id(id)))
-		en_users.discard(id)
+		logging.info("Spanish is the language selected by " + str(hide_id(chat_id)))
+		context.chat_data["language"] = 0
 		us.add_language(0)
-		await context.bot.send_message(chat_id=id, text=msg.get_message("language3", get_language(id)), parse_mode=ParseMode.HTML)
+		await context.bot.send_message(chat_id=chat_id, text=msg.get_message("language3", get_language(context)), parse_mode=ParseMode.HTML)
 
-#Handling clicks on InlineKeyboardButtons...
-async def button_click(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
+#Handling default clicks on InlineKeyboardButtons...
+async def default_button_click(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
 	query = update.callback_query
-	query.answer()
+	await query.answer()
 	if query.data.startswith("l"):
 		await set_language(update, context, query.data)
-	else:
-		await logging.info("Strange query from button recieved!")
 
+#Handling conversation clicks on InlineKeyboardButtons...
+async def conversation_button_click(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
+	chat_id = update.effective_chat.id
+	query = update.callback_query
+	await query.answer()
+	if query.data.startswith("a"):
+		if dt.now() - context.chat_data["admin_t"] < timedelta(minutes=5):
+			selection = int(query.data.split("_")[1])
+			if selection == 0:
+				await bot_usage(update, context)
+			return ConversationHandler.END
+		else:
+			chat_id = update.effective_chat.id
+			await context.bot.send_message(chat_id=chat_id, text=msg.get_message("admin_end", get_language(chat_id)), parse_mode=ParseMode.HTML)
+			return ConversationHandler.END
+	
 #Sending usage data...
 async def bot_usage(update, context):
-	id = update.effective_chat.id
-	m = update.message.text.split(" ")
-	if len(m) > 1 and m[1] == config["password"]:
-		m = us.build_usage_message()
-		await context.bot.send_message(chat_id=id, text=m, parse_mode=ParseMode.HTML)
-	else:
-		logging.info(hide_id(id) + " wanted to check bot usage data...")
-		await context.bot.send_message(chat_id=id, text=msg.get_message("intruder", get_language(id)), parse_mode=ParseMode.HTML)
-
-#Saving usage data...
-async def save_usage(update, context):
-	id = update.effective_chat.id
-	m = update.message.text.split(" ")
-	if len(m) > 1 and m[1] == config["password"]:
-		us.save_usage()
-		await context.bot.send_message(chat_id=id, text="¡Datos guardados!", parse_mode=ParseMode.HTML)
-	else:
-		logging.info(hide_id(id) + " wanted to save bot usage data...")
-		await context.bot.send_message(chat_id=id, text=msg.get_message("intruder", get_language(id)), parse_mode=ParseMode.HTML)
+	chat_id = update.effective_chat.id
+	m = us.build_usage_message()
+	await context.bot.send_message(chat_id=chat_id, text=m, parse_mode=ParseMode.HTML)
 
 #Notifying the user about out of context conversation...
 async def out_of_context(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
-	id = update.effective_chat.id
+	chat_id = update.effective_chat.id
 	us.add_outofcontext()
-	logging.info(str(hide_id(id)) + " sent out of context message...")
-	await context.bot.send_message(chat_id=id, text=msg.get_outofcontext(get_language(id)), parse_mode=ParseMode.HTML)
+	logging.info(str(hide_id(chat_id)) + " sent out of context message...")
+	await context.bot.send_message(chat_id=chat_id, text=msg.get_outofcontext(get_language(context)), parse_mode=ParseMode.HTML)
 
 #Sending error notification to administrator...
 async def error_notification(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
-	id = update.effective_chat.id
-	m = "An error ocurred! While comunicating with chat " + str(hide_id(id))
+	chat_id = update.effective_chat.id
+	m = "An error ocurred! While comunicating with chat " + str(hide_id(chat_id))
 	logging.info(m)
 	await context.bot.send_message(chat_id=config["admin_id"], text=m, parse_mode=ParseMode.HTML)
 
 #Hiding the first numbers of a chat id for the log...
-def hide_id(id):
-	s = str(id)
+def hide_id(chat_id):
+	s = str(chat_id)
 	return "****" + s[len(s)-4:]
 
 #Building the general conversation handler...
 def build_general_conversation_handler():
-	print("Building conversation handler...", end="\n")
+	print("Building general conversation handler...", end="\n")
 	handler = ConversationHandler(
-		entry_points=[CommandHandler("pcs", trigger_pcs),CommandHandler("chain", trigger_chain),
-					CommandHandler("error", trigger_error_submit)],
+		entry_points=[CommandHandler("start", start), CommandHandler("pcs", trigger_pcs),
+					CommandHandler("chain", trigger_chain), CommandHandler("random", trigger_random),
+					CommandHandler("admin", trigger_admin), CommandHandler("error", trigger_error_submit)],
 		states={
+			START_AN: [MessageHandler(filters.TEXT & ~filters.COMMAND, save_artistic_name)],
+			START_TS: [MessageHandler(filters.TEXT & ~filters.COMMAND, save_prefered_time_signature)],
 			PCS_S: [MessageHandler(filters.TEXT & ~filters.COMMAND, get_pcs_info)],
 			CHAIN_S: [MessageHandler(filters.TEXT & ~filters.COMMAND, get_new_chain)],
+			RANDOM_S: [MessageHandler(filters.TEXT & ~filters.COMMAND, set_random_size)],
+			RANDOM_D: [MessageHandler(filters.TEXT & ~filters.COMMAND, set_dice_size)],
+			RANDOM_B: [MessageHandler(filters.TEXT & ~filters.COMMAND, shuffle_message),
+								CallbackQueryHandler(random_conversation_button_click)],
+			RANDOM_C: [MessageHandler(filters.TEXT & ~filters.COMMAND, get_controlled_random)],
+			ADMIN: [CallbackQueryHandler(conversation_button_click)],
 			ERROR_1: [MessageHandler(filters.TEXT, report_command)],
 			ERROR_2: [MessageHandler(filters.TEXT & ~filters.COMMAND, report_error)],
 		},
 		fallbacks=[MessageHandler(filters.COMMAND, end_conversation)]
 		)
 	return handler
+
+#Building the general conversation handler...
+#def build_music_conversation_handler():
+#	print("Building general conversation handler...", end="\n")
+#	handler = ConversationHandler(
+#		entry_points=[CommandHandler("mus", trigger_mus)],
+#		states={
+#			CP_T: [MessageHandler(filters.TEXT & ~filters.COMMAND, get_mus_title)],
+#			CP_D: [MessageHandler(filters.TEXT, build_counterpoint)],
+#			CP_M: [MessageHandler(filters.TEXT & ~filters.COMMAND, get_mus_mode)],
+#		},
+#		fallbacks=[MessageHandler(filters.COMMAND, end_conversation)]
+#		)
+#	return handler
 
 #Here the magic happens...
 def main() -> None:
@@ -229,14 +450,13 @@ def main() -> None:
 		logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
 	print("Ready to build the bot...", end="\n")
 	app = Application.builder().token(config["token"]).build()
-	app.add_error_handler(error_notification)
-	app.add_handler(CommandHandler("start", start), group=2)
+	#app.add_error_handler(error_notification)
 	app.add_handler(CommandHandler("language", select_language), group=2)
 	app.add_handler(CommandHandler("help", print_help), group=2)
-	app.add_handler(CommandHandler("botusage", bot_usage), group=2)
-	app.add_handler(CommandHandler("saveusage", save_usage), group=2)
-	app.add_handler(CallbackQueryHandler(button_click), group=2)
+	app.add_handler(CommandHandler("privacy", print_privacy), group=2)
+	app.add_handler(CallbackQueryHandler(default_button_click), group=2)
 	app.add_handler(build_general_conversation_handler(), group=1)
+	#app.add_handler(build_music_conversation_handler(), group=1)
 	app.add_handler(MessageHandler(filters.TEXT & ~filters.COMMAND, out_of_context), group=1)
 	if config["webhook"]:
 		print("Ready to set webhook...", end="\n")